refactor: split world router
