+++ server/world/index.js
@@ -1,0 +1,11 @@
+// server/world/index.js
+import { Router } from 'express';
+import charactersRouter from './characters.js';
+import contextRouter from './context.js';
+
+const router = Router();
+
+router.use('/', charactersRouter);
+router.use('/', contextRouter);
+
+export default router;
